Move Fragment Classes to correct package
