+++ app/src/main/java/ca/ualberta/CMPUT3012019T02/alexandria/fragment/MessagesFragment.java
@@ -1,4 +1,4 @@
-package ca.ualberta.CMPUT3012019T02.alexandria.activity;
+package ca.ualberta.CMPUT3012019T02.alexandria.fragment;
 
 import android.os.Bundle;
 import android.support.annotation.NonNull;
@@ -10,12 +10,12 @@ import android.view.ViewGroup;
 
 import ca.ualberta.CMPUT3012019T02.alexandria.R;
 
-public class ExchangeFragment extends Fragment {
+public class MessagesFragment extends Fragment {
 
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater,
                              @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.fragment_exchange,null);
+        return inflater.inflate(R.layout.fragment_messages,null);
     }
 }
